refactoring
major refactoring. Removed a whole bunch of redundant code.
