+++ nbproject/private/private.xml
@@ -3,10 +3,27 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/Z:/git/GitPortable/Data/home/CPSO/test/cpso_r_k/CPSO_R_kTest.java</file>
-            <file>file:/Z:/git/GitPortable/Data/home/CPSO/src/cpso_r_k/CPSO_R_k.java</file>
-            <file>file:/Z:/git/GitPortable/Data/home/CPSO/src/cpso_h_k/CPSO_H_k.java</file>
-            <file>file:/Z:/git/GitPortable/Data/home/CPSO/src/cpso_r_k/Swarm.java</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/src/cpso_s_k/CPSO_S_k.java</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/test/cpso_s/Cpso_sSuite.java</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/test/cpso/ParticleTest.java</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/test/cpso/CpsoSuite.java</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/src/cpso_h_k/CPSO_H_k.java</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/test/RootSuite.java</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/test/cpso_s_k/Cpso_s_kSuite.java</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/test/cpso_h_k/Cpso_h_kSuite.java</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/test/cpso_r_k/CPSO_R_kTest.java</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/test/cpso_s/MainTest.java</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/test/cpso/SwarmTest.java</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/src/cpso/Swarm.java</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/src/cpso_r_k/CPSO_R_k.java</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/src/cpso_s/CPSO_S.java</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/test/cpso_s/CPSO_STest.java</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/test/cpso_h_k/CPSO_H_kTest.java</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/src/cpso_h_k/Main.java</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/src/cpso/Particle.java</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/test/cpso_s_k/MainTest.java</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/test/cpso_s_k/CPSO_S_kTest.java</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/src/cpso/CPSO.java</file>
         </group>
     </open-files>
 </project-private>
