+++ src/cpso_r_k/CPSO_R_k.java
@@ -5,219 +5,103 @@
  */
 package cpso_r_k;
 
+import cpso.*;
+
 
 /**
  *
  * @author Peter
  */
-public class CPSO_R_k {    
+public class CPSO_R_k extends CPSO {    
     
-        int loops;
-        boolean min = true;
-        private Swarm[] swarms; 
-        private Swarm pso_swarm;
-        private double[] solution;
-        int dimensionSize;
-        int swarmSize;
-        double C1 = 0.5;
-        double C2 = 0.3;
-        double INERTIA = 0.3;
-        double PSO_C1 = 0.5;
-        double PSO_C2 = 0.3;
-        double PSO_INTERTIA = 0.3;
-        int maxLoops;
-        int k = 0;
-        
-        /**
-         * Create a CPSO_R_k
-         * @param dimensionSize the overall dimension size
-         * @param maxLoops the max loops to perform
-         * @param swarmSize the number of particles in each swarm
-         * @param Inertia
-         * @param c1
-         * @param c2
-         * @param k the number of swarms
-         */
-        public CPSO_R_k(int dimensionSize, int maxLoops, int swarmSize, double Inertia, double c1, double c2, int k)
-        {
-            this.dimensionSize = dimensionSize;
-            this.maxLoops = maxLoops;
-            this.swarmSize = swarmSize;
-            this.INERTIA = Inertia;
-            this.C1 = c1;
-            this.C2 = c2;
-            if(k > dimensionSize) k = dimensionSize;
-            this.k = k;
-            InitializeSwarms();
-        }
-        
-        public void InitializeSwarms()
-        {
-            int remaining = 0;
-            int count = 0;
-            swarms = new Swarm[k];
-            pso_swarm = new Swarm(swarmSize, PSO_C1, PSO_C2, PSO_INTERTIA, min, dimensionSize);
-            solution = new double[dimensionSize];
-            for(int i = 0; i < k; i++)
-            {
-                //select the random grouping into k subgroups
-                int randomValue = (int)((Math.random())*((dimensionSize/k)-1))+1;                
-                int randomWidth = remaining + randomValue;
-                remaining = (int)(Math.ceil((dimensionSize/k)-randomValue));
-                swarms[i] = new Swarm(swarmSize, C1, C2, INERTIA, min, randomWidth);
-                
-                for(int j = 0; j < randomWidth; j++)
-                {
-                    solution[count++] = swarms[i].getParticles()[0].getPosition()[j];
-                }
-            }
-        }
-        
-        //calculate the fitness of the PSO
-        public void start()
-        {
-            for(int i = 0; i < maxLoops; i++)
-            {
-                
-                // <editor-fold desc="CPSO swarm">
-                    /////////////////////////////////////////
-                    /////    Update the CPSO Swarms    //////
-                    /////////////////////////////////////////
-                for (int s = 0; s < swarms.length; s++) //iterate through swarms
-                {                    
-                    for(Particle p : swarms[s].getParticles()){ //for each particle
-                        
-                        double fitness = CalculateFitness(s, p.getPosition()); //calculate the new fitness
-                        UpdateBests(fitness, p, swarms[s]);   
-                    }
-                    
-                    for (Particle p : swarms[s].getParticles()) //move the particles
-                    {
-                        swarms[s].UpdateVelocity(p);
-                        swarms[s].UpdatePosition(p);
-                    }                       
-                }
-                // </editor-fold>
-                
-                UpdateSolution();
-                
-            }
-            
-            for(int i = 0; i < solution.length; i++) //loop to print off solution
-            {
-                writeOutput("Solution "+(i+1)+": "+ solution[i]);
-            }
-            writeOutput("The final fitness value is: "+ CalculateFinalFitness());
-        }
-
-        /**
-         * Takes the new fitness and particle and determine if it is the new 
-         * global best and or personal best
-         * @param fitness the new fitness value
-         * @param p the particle that is being updated
-         * @param swarm the swarm that particle is from
-         */
-        private void UpdateBests(double fitness, Particle p, Swarm swarm)
-        {
-            if (p.UpdatePersonalBest(fitness, p.getPosition(), min))  //update the personal best
-                writeOutput("New Personal best for " + p + ": x=" + p.getPosition());
-
-            if ((swarm.getGlobalBest() == null) ||
-                (p.getFitness() < swarm.getGlobalBest().getFitness() && min) ||
-                (swarm.getGlobalBest().getFitness() < p.getFitness() && !min))      //update the global best
-            {
-                swarm.setGlobalBest(p);
-                writeOutput("New Global Best for Swarm " + swarm + ": x=" + p.getPosition());
-            }
-        }
-        
-         /**
-         * Update to the best current solution by taking the global best values
-         */
-        private void UpdateSolution()
-        {
-            int index = 0;
-            for(int i = 0; i < swarms.length; i++)
-            {
-                Particle best = swarms[i].getGlobalBest();
-                if(best == null){
-                    index+= swarms[i].getParticles()[0].getPosition().length;
-                    continue;
-                }
-                else{
-                    for(int j = 0; j < best.getPosition().length; j++)
-                    {
-                        solution[index++] = best.getPosition()[j];
-                    }
-                }
-            }
-        }
-        
-        /**
-         * Calculate the fitness of the current swarm
-         * @param position the current position
-         * @return 
-         */
-        public double CalculateFitness(int index, double[] position)
-        {
-            double fitness = 0;
-            int count = 0;
-            for(int i = 0; i < k; i++)
-            {
-                if(i == index)
-                    for(int j = 0; j < swarms[i].getParticles()[0].getPosition().length; j++)
-                    {
-                        fitness += Math.log(position[j]);
-                        count++;
-                    }
-                else
-                    for(int j = 0; j < swarms[i].getParticles()[0].getPosition().length; j++)
-                        fitness += Math.log(getSolution()[count++]);
-            }
-            return fitness;   
-        }
-        
-        /**
-         * Returns the final fitness based on the solution function
-         * @return fitness
-         */
-        public double CalculateFinalFitness()
-        {
-            double fitness = 0;
-            for(int i = 0; i < dimensionSize; i++)
-            {
-                fitness += Math.log(getSolution()[i]);
-            }
-            return fitness; 
-        }
-
-        private void writeOutput(String output)
-        {
-            System.out.println(output);
-        }
-
-        /**
-         * @return the swarms
-         */
-        public Swarm[] getSwarms() {
-            return swarms;
-        }
+    int loops;
+    boolean min = true;
+    private Swarm[] swarms; 
+    private double[] solution;
+    int dimensionSize;
+    int swarmSize;
+    double C1 = 0.5;
+    double C2 = 0.3;
+    double INERTIA = 0.3;
+    double PSO_C1 = 0.5;
+    double PSO_C2 = 0.3;
+    double PSO_INTERTIA = 0.3;
+    int maxLoops;
+    int k = 0;
 
     /**
-     * @return the solution
+     * Create a CPSO_R_k
+     * @param dimensionSize the overall dimension size
+     * @param maxLoops the max loops to perform
+     * @param swarmSize the number of particles in each swarm
+     * @param Inertia
+     * @param c1
+     * @param c2
+     * @param k the number of swarms
      */
-    public double[] getSolution() {
-        return solution;
+    public CPSO_R_k(int dimensionSize, int maxLoops, int swarmSize, double Inertia, double c1, double c2, int k)
+    {
+        super(dimensionSize, maxLoops, swarmSize, Inertia, c1, c2, k);
+        this.InitializeSwarms();
     }
 
-    /**
-     * @param solution the solution to set
-     */
-    public void setSolution(double[] solution) throws Exception {
-        if(solution.length == dimensionSize)
+    @Override
+    public void InitializeSwarms()
+    {
+        int remaining = 0;
+        int count = 0;
+        swarms = new Swarm[k];
+        solution = new double[dimensionSize];
+        for(int i = 0; i < k; i++)
         {
-            this.solution = solution;
+            //select the random grouping into k subgroups
+            int randomValue = (int)((Math.random())*((dimensionSize/k)-1))+1;                
+            int randomWidth = remaining + randomValue;
+            remaining = (int)(Math.ceil((dimensionSize/k)-randomValue));
+            swarms[i] = new Swarm(swarmSize, C1, C2, INERTIA, min, randomWidth);
+
+            for(int j = 0; j < randomWidth; j++)
+            {
+                solution[count++] = swarms[i].getParticles()[0].getPosition()[j];
+            }
         }
-        else throw new Exception("Incorrect solution size");
+    }
+
+    //calculate the fitness of the PSO
+    @Override
+    public void start()
+    {
+        for(int i = 0; i < maxLoops; i++)
+        {
+
+            // <editor-fold desc="CPSO swarm">
+                /////////////////////////////////////////
+                /////    Update the CPSO Swarms    //////
+                /////////////////////////////////////////
+            for (int s = 0; s < swarms.length; s++) //iterate through swarms
+            {                    
+                for(Particle p : swarms[s].getParticles()){ //for each particle
+
+                    double fitness = CalculateFitness(s, p.getPosition()); //calculate the new fitness
+                    UpdateBests(fitness, p, swarms[s]);   
+                }
+
+                for (Particle p : swarms[s].getParticles()) //move the particles
+                {
+                    swarms[s].UpdateVelocity(p);
+                    swarms[s].UpdatePosition(p);
+                }                       
+            }
+            // </editor-fold>
+
+            UpdateSolution();
+
+        }
+
+        for(int i = 0; i < solution.length; i++) //loop to print off solution
+        {
+            writeOutput("Solution "+(i+1)+": "+ solution[i]);
+        }
+        writeOutput("The final fitness value is: "+ CalculateFinalFitness());
     }
 }
