+++ src/cpso_s/CPSO_S.java
@@ -4,12 +4,13 @@
  * and open the template in the editor.
  */
 package cpso_s;
+import cpso.*;
 
 /**
  *
  * @author Peter
  */
-public class CPSO_S {    
+public class CPSO_S extends CPSO {    
     
         int loops;
         boolean min = true;
@@ -21,31 +22,26 @@ public class CPSO_S {
         double C2 = 0.3;
         double INERTIA = 0.3;
         int maxLoops;
+        int k;
         
-        public CPSO_S(int dimensionSize, int maxLoops, int swarmSize, double Inertia, double c1, double c2)
+        public CPSO_S(int dimensionSize, int maxLoops, int swarmSize, double Inertia, double c1, double c2, int k)
         {
+            super();
             this.dimensionSize = dimensionSize;
             this.maxLoops = maxLoops;
             this.swarmSize = swarmSize;
             this.INERTIA = Inertia;
             this.C1 = c1;
             this.C2 = c2;
-            InitializeSwarms();
-        }
-        
-        public void InitializeSwarms()
-        {
-            swarms = new Swarm[dimensionSize];
+            if(k > dimensionSize) k = dimensionSize;
+            this.k = k;            
             solution = new double[dimensionSize];
-            for(int i = 0; i < dimensionSize; i++)
-            {
-                swarms[i] = new Swarm(swarmSize, C1, C2, INERTIA, min);
-                
-                solution[i] = swarms[i].getParticles()[0].getPosition();
-            }
+            swarms = InitializeSwarms(k, getSolution(), dimensionSize, swarmSize,
+                            C1, C2, Inertia, min);
         }
         
         //calculate the fitness of the PSO
+        @Override
         public void start()
         {
             for(int i = 0; i < maxLoops; i++)
@@ -54,19 +50,8 @@ public class CPSO_S {
                 {
                     for(Particle p : swarms[s].getParticles()){ //for each particle
                         
-                        double fitness = CalculateFitness(s, p.getPosition()); //calculate the new fitness
-                        
-                        if (p.UpdatePersonalBest(fitness, p.getPosition(), min))  //update the personal best
-                            writeOutput("New Personal best for " + p + ": x=" + p.getPosition());
-                        
-                        if ((swarms[s].getGlobalBest() == null) ||
-                            (p.getFitness() < swarms[s].getGlobalBest().getFitness() && min) ||
-                            (swarms[s].getGlobalBest().getFitness() < p.getFitness() && !min))      //update the global best
-                        {
-
-                            swarms[s].setGlobalBest(p);
-                            writeOutput("New Global Best for Swarm " + s + ": x=" + p.getPosition());
-                        }
+                        double fitness = CalculateFitness(s, p.getPosition(), k, swarms, getSolution()); //calculate the new fitness
+                        UpdateBests(fitness, p, swarms[s], min); 
                     }
                     
                     for (Particle p : swarms[s].getParticles()) //move the particles
@@ -76,72 +61,24 @@ public class CPSO_S {
                     } 
                     
                 }
-                    UpdateSolution();
+                    UpdateSolution(swarms, solution);
             }
             
             for(int i = 0; i < solution.length; i++) //loop to print off solution
             {
                 writeOutput("Solution "+(i+1)+": "+ solution[i]);
             }
-            writeOutput("The final fitness value is: "+ CalculateFinalFitness());
-        }
-
-        /**
-         * Calculate the fitness of the current swarm
-         * @param position the current position
-         * @return 
-         */
-        public double CalculateFitness(int index, double position)
-        {
-            double fitness = 0;
-            for(int i = 0; i < getSolution().length; i++)
-            {
-                if(i == index)
-                    fitness += Math.log(position);
-                else
-                    fitness += Math.log(getSolution()[i]);
-            }
-            return fitness;   
+            writeOutput("The final fitness value is: "+ CalculateFinalFitness(dimensionSize, getSolution()));
         }
         
-        /**
-         * Returns the final fitness based on the solution function
-         * @return fitness
-         */
-        public double CalculateFinalFitness()
-        {
-            double fitness = 0;
-            for(int i = 0; i < getSolution().length; i++)
-            {
-                fitness += Math.log(getSolution()[i]);
-            }
-            return fitness; 
-        }
         
-        /**
-         * Update to the best current solution by taking the global best values
-         */
-        private void UpdateSolution()
-        {
-            for(int i = 0; i < swarms.length; i++)
-            {
-                Particle best = swarms[i].getGlobalBest();
-                if(best == null) continue;
-                else solution[i] = best.getPosition();
-            }
-        }
 
-        private void writeOutput(String output)
-        {
-            System.out.println(output);
-        }
-
-        /**
-         * @return the swarms
-         */
-        public Swarm[] getSwarms() {
-            return swarms;
-        }
+    /**
+     * @return the swarms
+     */
+    public Swarm[] getSwarms() {
+        return swarms;
+    }
 
     /**
      * @return the solution
@@ -152,6 +89,7 @@ public class CPSO_S {
 
     /**
      * @param solution the solution to set
+     * @throws exception incorrect solution size
      */
     public void setSolution(double[] solution) throws Exception {
         if(solution.length == dimensionSize)
