+++ src/cpso_s/Main.java
@@ -17,7 +17,7 @@ public class Main {
     public static void main(String[] args) {
         // TODO code application logic here
         int dimensionSize = 5;
-        CPSO_S cpso = new CPSO_S(dimensionSize, 200, 20, 0.5, 0.3, 0.2);
+        CPSO_S cpso = new CPSO_S(dimensionSize, 200, 20, 0.5, 0.3, 0.2, 1);
         cpso.start();
     }
