+++ src/cpso_s_k/CPSO_S_k.java
@@ -4,176 +4,59 @@
  * and open the template in the editor.
  */
 package cpso_s_k;
+import cpso.*;
 
 /**
  *
  * @author Peter
  */
-public class CPSO_S_k {    
+public class CPSO_S_k extends CPSO {    
     
-        int loops;
-        boolean min = true;
-        private Swarm[] swarms; 
-        private double[] solution;
-        int dimensionSize;
-        int swarmSize;
-        double C1 = 0.5;
-        double C2 = 0.3;
-        double INERTIA = 0.3;
-        int maxLoops;
-        int k = 0;
-        
-        public CPSO_S_k(int dimensionSize, int maxLoops, int swarmSize, double Inertia, double c1, double c2, int k)
-        {
-            this.dimensionSize = dimensionSize;
-            this.maxLoops = maxLoops;
-            this.swarmSize = swarmSize;
-            this.INERTIA = Inertia;
-            this.C1 = c1;
-            this.C2 = c2;
-            this.k = k;
-            InitializeSwarms();
-        }
-        
-        public void InitializeSwarms()
-        {
-            swarms = new Swarm[dimensionSize/k];
-            solution = new double[dimensionSize];
-            for(int i = 0; i < dimensionSize/k; i++)
-            {
-                swarms[i] = new Swarm(swarmSize, C1, C2, INERTIA, min, k);
-                
-                for(int j = 0; j < k; j++)
-                {
-                    solution[(i*k)+j] = swarms[i].getParticles()[0].getPosition()[j];
-                }
-            }
-        }
-        
-        //calculate the fitness of the PSO
-        public void start()
-        {
-            for(int i = 0; i < maxLoops; i++)
-            {
-                for (int s = 0; s < swarms.length; s++) //iterate through swarms
-                {
-                    for(Particle p : swarms[s].getParticles()){ //for each particle
-                        
-                        double fitness = CalculateFitness(s, p.getPosition()); //calculate the new fitness
-                        
-                        if (p.UpdatePersonalBest(fitness, p.getPosition(), min))  //update the personal best
-                            writeOutput("New Personal best for " + p + ": x=" + p.getPosition());
-                        
-                        if ((swarms[s].getGlobalBest() == null) ||
-                            (p.getFitness() < swarms[s].getGlobalBest().getFitness() && min) ||
-                            (swarms[s].getGlobalBest().getFitness() < p.getFitness() && !min))      //update the global best
-                        {
+    int loops;
+    boolean min = true;
+    private Swarm[] swarms; 
+    private double[] solution;
+    int dimensionSize;
+    int swarmSize;
+    double C1 = 0.5;
+    double C2 = 0.3;
+    double INERTIA = 0.3;
+    int maxLoops;
+    int k = 0;
 
-                            swarms[s].setGlobalBest(p);
-                            writeOutput("New Global Best for Swarm " + s + ": x=" + p.getPosition());
-                        }
-                    }
-                    
-                    for (Particle p : swarms[s].getParticles()) //move the particles
-                    {
-                        swarms[s].UpdateVelocity(p);
-                        swarms[s].UpdatePosition(p);
-                    }                       
-                }
-                UpdateSolution();
-            }
-            
-            for(int i = 0; i < solution.length; i++) //loop to print off solution
-            {
-                writeOutput("Solution "+(i+1)+": "+ solution[i]);
-            }
-            writeOutput("The final fitness value is: "+ CalculateFinalFitness());
-        }
-
-        /**
-         * Calculate the fitness of the current swarm
-         * @param position the current position
-         * @return 
-         */
-        public double CalculateFitness(int index, double[] position)
-        {
-            double fitness = 0;
-            for(int i = 0; i < (getSolution().length/k); i++)
-            {
-                if(i == index)
-                    for(int j = 0; j < k; j++)
-                        fitness += Math.log(position[j]);
-                else
-                    for(int j = 0; j < k; j++)
-                        fitness += Math.log(getSolution()[(i*k)+j]);
-            }
-            return fitness;   
-        }
-        
-         /**
-         * Update to the best current solution by taking the global best values
-         */
-        private void UpdateSolution()
-        {
-            int index = 0;
-            for(int i = 0; i < swarms.length; i++)
-            {
-                Particle best = swarms[i].getGlobalBest();
-                if(best == null){
-                    index+= k;
-                    continue;
-                }
-                else{
-                    for(int j = 0; j < best.getPosition().length; j++)
-                    {
-                        solution[index] = best.getPosition()[j];
-                        index++;
-                    }
-                }
-            }
-        }
-        
-        /**
-         * Returns the final fitness based on the solution function
-         * @return fitness
-         */
-        public double CalculateFinalFitness()
-        {
-            double fitness = 0;
-            for(int i = 0; i < (getSolution().length/k); i++)
-            {
-                fitness += Math.log(getSolution()[i]);
-            }
-            return fitness; 
-        }
-
-        private void writeOutput(String output)
-        {
-            System.out.println(output);
-        }
-
-        /**
-         * @return the swarms
-         */
-        public Swarm[] getSwarms() {
-            return swarms;
-        }
-
-    /**
-     * @return the solution
-     */
-    public double[] getSolution() {
-        return solution;
+    public CPSO_S_k(int dimensionSize, int maxLoops, int swarmSize, double Inertia, double c1, double c2, int k)
+    {
+        super(dimensionSize, maxLoops, swarmSize, Inertia, c1, c2, k);
+        super.InitializeSwarms();
     }
 
-    /**
-     * @param solution the solution to set
-     */
-    public void setSolution(double[] solution) throws Exception {
-        if(solution.length == dimensionSize)
+    //calculate the fitness of the PSO
+    @Override
+    public void start()
+    {
+        for(int i = 0; i < maxLoops; i++)
         {
-            this.solution = solution;
+            for (int s = 0; s < swarms.length; s++) //iterate through swarms
+            {
+                for(Particle p : swarms[s].getParticles()){ //for each particle
+
+                    double fitness = CalculateFitness(s, p.getPosition()); //calculate the new fitness
+                    UpdateBests(fitness, p, swarms[s]);  
+                }
+
+                for (Particle p : swarms[s].getParticles()) //move the particles
+                {
+                    swarms[s].UpdateVelocity(p);
+                    swarms[s].UpdatePosition(p);
+                }                       
+            }
+            UpdateSolution();
         }
-        else throw new Exception("Incorrect solution size");
+
+        for(int i = 0; i < solution.length; i++) //loop to print off solution
+        {
+            writeOutput("Solution "+(i+1)+": "+ solution[i]);
+        }
+        writeOutput("The final fitness value is: "+ CalculateFinalFitness());
     }
 }
