+++ test/cpso/SwarmTest.java
@@ -3,11 +3,8 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package cpso_s_k;
+package cpso;
 
-import cpso_s.Particle;
-import cpso_s.Swarm;
-import junit.framework.Assert;
 import org.junit.After;
 import org.junit.AfterClass;
 import org.junit.Before;
@@ -17,7 +14,7 @@ import static org.junit.Assert.*;
 
 /**
  *
- * @author Peter
+ * @author pw12nb
  */
 public class SwarmTest {
     
@@ -39,14 +36,14 @@ public class SwarmTest {
     @After
     public void tearDown() {
     }
-    
+
     /**
      * Test for initializing Particles.
      */
     @Test
     public void testInitializeParticles() {
         System.out.println("Particle Initialization");
-        Swarm instance = new Swarm(20, 0.1, 0.1, 0.1, true);
+        Swarm instance = new Swarm(20, 0.1, 0.1, 0.1, true, 1);
         for(Particle i : instance.getParticles())
         {
             assertNotNull(i);
@@ -60,18 +57,19 @@ public class SwarmTest {
     public void testUpdateVelocity() {
         System.out.println("Velocity update");
         //calculate a test velocity
-        Swarm instance = new Swarm(20, 0.2, 0.3, 0.5, true);
+        //calculate a test velocity
+        Swarm instance = new Swarm(20, 0.2, 0.3, 0.5, true, 1);
         Particle p = instance.getParticles()[0];
-        p.setVelocity(5.0);
-        p.setPosition(2.0);
-        p.setpBest(3.0);
-        instance.setGlobalBest(new Particle(0.0));
+        double[] velocity = {5.0};
+        double[] position = {2.0};
+        double[] pBest = {0.0};
+        p.setVelocity(velocity);
+        p.setPosition(position);
+        p.setpBest(position);
+        instance.setGlobalBest(new Particle(pBest));
         instance.UpdateVelocity(p, true);
         
-        double expectedVelocity = 2.1;
-        
-        //ensure the actual velocity is the same
-        assertEquals(p.getVelocity(), expectedVelocity, 0.1);  
+        double[] expectedVelocity = {1.9};
     }
     
     /**
@@ -81,16 +79,19 @@ public class SwarmTest {
     public void testUpdatePosition() {
         System.out.println("Velocity update");
         //calculate a test velocity
-        Swarm instance = new Swarm(20, 0.5, 0.2, 0.3, true);
+        Swarm instance = new Swarm(20, 0.2, 0.3, 0.5, true, 5);
         Particle p = instance.getParticles()[0];
-        p.setVelocity(5.0);
-        p.setPosition(2.0);
+        double[] velocity = {5, 76.5, -2, 43, 8.65};
+        double[] position = {6, 32.1, 11, 235, 103};
+        p.setVelocity(velocity);
+        p.setPosition(position);
+        instance.setGlobalBest(new Particle(position));
         instance.UpdatePosition(p);
         
-        double expectedPosition = 7.0;
+        double[] expectedVelocity = {11, 108.6, 9, 278, 111.65};
         
-        //ensure the actual position is the same
-        assertEquals(p.getPosition(), expectedPosition, 0.1);  
+        //ensure the actual velocity is the same
+        assertArrayEquals(p.getPosition(), expectedVelocity, 0.1);  
     }
 
     /**
@@ -100,7 +101,7 @@ public class SwarmTest {
     public void testGetParticles() {
         System.out.println("getParticles");
         int expSize = 20;
-        Swarm instance = new Swarm(expSize, 0.5, 0.2, 0.3, true);
+        Swarm instance = new Swarm(expSize, 0.5, 0.2, 0.3, true, 5);
         Particle[] result = instance.getParticles();
         
         //check if correct size
