+++ test/cpso_h_k/Cpso_h_kSuite.java
@@ -17,7 +17,7 @@ import org.junit.runners.Suite;
  * @author Peter
  */
 @RunWith(Suite.class)
-@Suite.SuiteClasses({CPSO_H_kTest.class, ParticleTest.class, MainTest.class, SwarmTest.class})
+@Suite.SuiteClasses({CPSO_H_kTest.class, MainTest.class})
 public class Cpso_h_kSuite {
 
     @BeforeClass
