+++ test/cpso_s/Cpso_sSuite.java
@@ -17,7 +17,7 @@ import org.junit.runners.Suite;
  * @author Peter
  */
 @RunWith(Suite.class)
-@Suite.SuiteClasses({CPSO_STest.class, ParticleTest.class, SwarmTest.class, MainTest.class})
+@Suite.SuiteClasses({CPSO_STest.class, MainTest.class})
 public class Cpso_sSuite {
 
     @BeforeClass
